+++ src/wurf/lock_cache.py
@@ -2,8 +2,10 @@
 # encoding: utf-8
 
 import json
+import hashlib
+import os
 
-from .error import WurfError, DependencyError
+from .error import WurfError
 
 
 class LockCache(object):
@@ -17,17 +19,12 @@ class LockCache(object):
         return self.lock_cache["dependencies"][dependency.name]["path"]
 
     def checkout(self, dependency):
+        assert dependency.resolver == "git"
         return self.lock_cache["dependencies"][dependency.name]["checkout"]
 
     def check_sha1(self, dependency):
-
         lock_sha1 = self.lock_cache["dependencies"][dependency.name]["sha1"]
-
-        if dependency.sha1 != lock_sha1:
-            raise DependencyError(
-                msg="SHA1 mismatch. Locked SHA1: {}".format(lock_sha1),
-                dependency=dependency,
-            )
+        return dependency.sha1 != lock_sha1
 
     def write_to_file(self, lock_path):
         with open(lock_path, "w") as lock_file:
@@ -44,6 +41,32 @@ class LockCache(object):
             "sha1": dependency.sha1,
             "checkout": checkout,
         }
+
+    def add_file(self, dependency, path):
+        self.lock_cache["dependencies"][dependency.name] = {
+            "sha1": dependency.sha1,
+            "file_hash": self.__calculate_file_hash(path),
+        }
+
+    def check_file_hash(self, dependency, path):
+        assert dependency.resolver == "http"
+        lock_file_hash = self.lock_cache["dependencies"][dependency.name]["file_hash"]
+        return lock_file_hash != self.__calculate_file_hash(path)
+
+    def __calculate_file_hash(self, path):
+        assert os.path.exists(path)
+        sha1 = hashlib.sha1()
+        if os.path.isfile(path):
+            sha1.update(open(path, "rb").read())
+        elif os.path.isdir(path):
+            # Calculate the hash of all files in the directory
+            for root, _, files in os.walk(path):
+                for file in files:
+                    f = os.path.join(root, file)
+                    sha1.update(open(f, "rb").read())
+        else:
+            raise WurfError("Unknown file type: {}".format(path))
+        return sha1.hexdigest()
 
     def __contains__(self, dependency):
         """